data_science/reddit_sentiment/slack/requirements.txt
psycopg2-binary
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
SQLAlchemy==1.4.44